UGENE-6991. Remove qint64 getContentIndentY from SequenceView renderers

+++ src/corelibs/U2View/src/ov_msa/view_rendering/McaReferenceAreaRenderer.cpp
@@ -48,7 +48,7 @@ McaReferenceAreaRenderer::McaReferenceAreaRenderer(PanView *panView, SequenceObj
     setFont(maEditor->getFont());
 }
 
-qint64 McaReferenceAreaRenderer::getMinimumHeight() const {
+int McaReferenceAreaRenderer::getMinimumHeight() const {
     return commonMetrics.lineHeight;
 }
